requirements.txt
# GitGallery - V1 Prototype
# Python 3.11+

# GUI
PySide6>=6.6.0

# GitHub API
PyGithub>=2.1.1
requests>=2.31.0

# Image processing (thumbnails)
Pillow>=10.0.0

# OAuth (device flow / local callback)
authlib>=1.2.1

# Optional: GitPython (alternative to Git CLI)
# GitPython>=3.1.40

# Development / tests
pytest>=7.4.0

python-dotenv
